+++ nengo_spinnaker_gfe/helpful_functions.py
@@ -2,10 +2,6 @@ from data_specification.enums import DataType
 import numpy
 
 from nengo_spinnaker_gfe import constants
-from nengo_spinnaker_gfe.nengo_exceptions import NotRecognisedFilterType
-from nengo_spinnaker_gfe.nengo_filters.linear_filter import LinearFilter
-from nengo_spinnaker_gfe.nengo_filters.low_pass_filter import LowPassFilter
-from nengo_spinnaker_gfe.nengo_filters.none_filter import NoneFilter
 
 
 def get_seed(nengo_object):
@@ -98,54 +94,12 @@ def sdram_size_in_bytes_for_filter_region(filters):
     return total
 
 
-def write_filter_region(
-        spec, machine_time_step_in_seconds, input_slice, filters):
-    # write how many filters there are
-    low_pass_filters, none_type_filters, linear_filters = \
-        _separate_filter_types(filters)
-
-    # process the low pass
-    spec.write_value(len(low_pass_filters))
-    for low_pass_filter in low_pass_filters:
-        low_pass_filter.write_spec(
-            spec, machine_time_step_in_seconds, input_slice.n_atoms)
-
-    # process none
-    spec.write_value(len(none_type_filters))
-    for none_type_filter in none_type_filters:
-        none_type_filter.write_spec(
-            spec, machine_time_step_in_seconds, input_slice.n_atoms)
-
-    # process linear
-    spec.write_value(len(linear_filters))
-    for linear_filter in linear_filters:
-        linear_filter.write_spec(
-            spec, machine_time_step_in_seconds, input_slice.n_atoms)
-
-
 def write_routing_region(spec, routing_info, machine_graph, vertex):
     for outgoing_partition in (
             machine_graph.get_outgoing_edge_partitions_starting_at_vertex(
                 vertex)):
         spec.write_value(
             routing_info.get_first_key_from_partition(outgoing_partition))
-
-
-def _separate_filter_types(filters):
-    low_pass = list()
-    none_type = list()
-    linear = list()
-    for filter_to_categorise in filters:
-        if isinstance(filter_to_categorise, NoneFilter):
-            none_type.append(filter_to_categorise)
-        elif isinstance(filter_to_categorise, LowPassFilter):
-            low_pass.append(filter_to_categorise)
-        elif isinstance(filter_to_categorise, LinearFilter):
-            linear.append(filter_to_categorise)
-        else:
-            raise NotRecognisedFilterType(
-                "don't know how to separate this filter")
-    return low_pass, none_type, linear
 
 
 def sdram_size_in_bytes_for_routing_region(n_keys):
