+++ nengo_spinnaker_gfe/machine_vertices/sdp_transmitter_machine_vertex.py
@@ -8,6 +8,7 @@ from nengo_spinnaker_gfe.abstracts.abstract_accepts_multicast_signals import \
     AbstractAcceptsMulticastSignals
 from nengo_spinnaker_gfe.abstracts.abstract_nengo_machine_vertex import \
     AbstractNengoMachineVertex
+from nengo_spinnaker_gfe.nengo_filters import filter_region_writer
 from pacman.executor.injection_decorator import inject_items
 from pacman.model.resources import ResourceContainer, SDRAMResource, \
     IPtagResource
@@ -140,7 +141,7 @@ class SDPTransmitterMachineVertex(
 
         # fill in filter region
         spec.switch_write_focus(self.DATA_REGIONS.FILTER.value)
-        helpful_functions.write_filter_region(
+        filter_region_writer.write_filter_region(
             spec, machine_time_step_in_seconds, self._input_slice,
             self._input_filters)
